refactor(test): move resultant matrix tests to dedicated file
Move TestSylvesterMatrix and TestBezoutMatrix from test_bezier.py to
test_bezier_resultant_matrices.py, following the existing pattern of
per-module test files.
